Add files via upload
This version will run much faster. Uses the method suggested from the game of life example on Python Arcade website. Instead of drawing and redrawing rectangles for each state update, we now use a sprite overlay and change the alpha value between 0 and 255 for corresponding locations to simulate "alive" and "dead" cells